Move all isobank commands from main bot script to a cog
